rename `shell_attr` to `shell_attributes.h`
