+++ sandman/source/shell.h
@@ -2,7 +2,7 @@
 
 #include "common/enum.h"
 #include "common/box.h"
-#include "shell_attr.h"
+#include "shell_attributes.h"
 
 #include <array>
 #include <cstddef>
