+++ sandman/source/shell_attributes.h
